requirements.txt
python-telegram-bot>=20,<22
python-dotenv>=1.0.0
python-dateutil>=2.9.0.post0
tzdata>=2024.1
pymongo>=4.6
fastapi>=0.110
pydantic>=2.7
uvicorn>=0.30
httpx>=0.27